Changed the name of the file and the class Agents for Customer since Agents is to general and maybe later we will include another kind of agents that should be on a different file and class.
Created a static stl vector in customer class to store the history of
attendance on every week for every instance(customer), this is because
every customer should take part of that information to make a decision.

TODO:
Documentation of what has been done so far since it seems to be
taking a stable course.
Write the decition functions.